git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@4450 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ annotation/StandAlone_AnnotationModule.sh
@@ -15,6 +15,11 @@ else
 	multi=$6
 	script_path=$7
 	## paths
+	if [ ! -f $file ]
+	then
+		echo "input file is not found"
+		exit 1;
+	fi	
 	mkdir -p $output
 	cp $tool_info $output/$sample.tool_info.txt
 	tool_info=$output/$sample.tool_info.txt
